4.12 fix some bugs

+++ SXK/app/src/main/java/com/example/cfwifine/sxk/Section/MineNC/Controller/RentAndSale/MineRent/Model/MineItemBoZhuWaitReceModel.java
@@ -75,7 +75,7 @@ public class MineItemBoZhuWaitReceModel {
         private int userid;
         private ReceiverBean receiver;
         private String message;
-        private int point;
+        private double point;
         private int commission;
         private int share;
         private String oddNumber;
@@ -158,11 +158,11 @@ public class MineItemBoZhuWaitReceModel {
             this.message = message;
         }
 
-        public int getPoint() {
+        public double getPoint() {
             return point;
         }
 
-        public void setPoint(int point) {
+        public void setPoint(double point) {
             this.point = point;
         }
